Day 9 finished

+++ Day 9/Node.h
@@ -2,15 +2,15 @@
 using namespace std;
 
 struct Node{
-    int number;
+    long long int number;
     Node* next;
     Node* last;
-    Node(int dat){
+    Node(long long int dat){
         number = dat;
     }
 };
 
-void insert(Node* before, Node* ins, int& size){
+void insert(Node* before, Node* ins, long long int& size){
     ins->next = before->next;
     ins->next->last = ins;
     ins->last = before;
@@ -18,14 +18,14 @@ void insert(Node* before, Node* ins, int& size){
     size++;
 }
 
-void insert(Node* before, int val, int& size){
+void insert(Node* before, long long int val, long long int& size){
     insert(before, new Node(val), size);
 }
 
-int remove(Node* rem, int& size){
+long long int remove(Node* rem, long long int& size){
     rem->last->next = rem->next;
     rem->next->last = rem->last;
-    int ret = rem->number;
+    long long int ret = rem->number;
     delete rem;
     size--;
     return (ret);
@@ -39,9 +39,9 @@ void left(Node*& p){
     p = p->last;
 }
 
-void print(Node* head, int size){
+void print(Node* head, long long int size){
     Node* p = head;
-    for(int i =0;i<size;i++){
+    for(long long int i =0;i<size;i++){
         cout<<p->number<<" ";
         right(p);
     }cout<<endl;
